refactor auth route
